refactor: move MetricsSnapshot type to @chapa/shared
- MetricsSnapshot and SnapshotPenalty now exported from @chapa/shared
- All imports updated to use shared package
- apps/web/lib/history/types.ts re-exports for backward compat

Fixes #318
